+++ src/imap_client.php
@@ -54,13 +54,25 @@ class ImapClient {
     }
 
 
+    public function load_one_email_fully($download_email_id, $user) {
+        if ($this->load_one_email($download_email_id, $user) !== null) {
+            $headers = imap_fetchheader($this->mailbox->getImapStream(), $download_email_id, FT_UID);
+            $body = imap_body($this->mailbox->getImapStream(), $download_email_id, FT_UID);
+            return $headers . "\n" . $body;
+        } else {
+            return null;
+        }
+
+    }
+
     /**
      * Load emails using the $mail_ids, the mails have to match the $address in TO or CC.
      * @param $mail_ids array of integer ids
      * @param $user User
      * @return array of emails
      */
-    private function _load_emails(array $mail_ids, User $user) {
+    private
+    function _load_emails(array $mail_ids, User $user) {
         $emails = array();
         foreach ($mail_ids as $id) {
             $mail = $this->mailbox->getMail($id);
@@ -75,7 +87,8 @@ class ImapClient {
     /**
      * deletes messages older than X days.
      */
-    public function delete_old_messages(string $delete_messages_older_than) {
+    public
+    function delete_old_messages(string $delete_messages_older_than) {
         $ids = $this->mailbox->searchMailbox('BEFORE ' . date('d-M-Y', strtotime($delete_messages_older_than)));
         foreach ($ids as $id) {
             $this->mailbox->deleteMail($id);